requirements.txt
pandas
pandasql
numpy
matplotlib
seaborn
scikit-learn
joblib
streamlit
plotly
sqlalchemy
psycopg2-binary
xgboost
missingno
